Remove verifyExpRelease
VerifyReleaseTask depending on doing a GitHub Release for every single build and checking it via the API. This is very wasteful and I am ending this practice. Technically we could tag every release when built and just check a tag exists, but this is also unnecessary - instead we just ensure we run all the tests every time we do a manually triggered release.